// src/api/generated/models/activityLogItemDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Flowforge API
 * The Flowforge workflow orchestration engine API documentation
 * OpenAPI spec version: 1.0
 */
import type { ActivityLogItemDtoMetadata } from './activityLogItemDtoMetadata';
import type { ActorInfoDto } from './actorInfoDto';

export interface ActivityLogItemDto {
  id: string;
  organizationId: string;
  action: string;
  targetType: string;
  targetId?: string;
  targetName?: string;
  /** Additional context for the action */
  metadata?: ActivityLogItemDtoMetadata;
  actor: ActorInfoDto;
  createdAt: string;
}